requirements.txt
smbus2==0.5.0
paho-mqtt==2.1.0
RPi.GPIO==0.7.1
python-dotenv==1.1.0
